[qca-nss-drv] Consume skbs based on their head
Skbs used to be consumed based on the last descriptor
information. Changed to consume on the first descriptor.

Consuming skbs on the last descriptor created issues when
exceptioning packets. In firmware, setting the exception src on
a sg chain only applies to the first skb. This is why we
now should except on the first skb to go to the approriate
callback.

Change-Id: Id33c880fc80e5013f60c328ea21fa07490b6db9d
Signed-off-by: Thomas Wu <wthomas@codeaurora.org>