[qca-nss-drv] Combine flush calls for descriptors
Calls to dmac_clean_range are expensive since they incur
a function call, arithmetic for calculating range, and a
dsb for sync inside the function. We could use inline assembly
to only do one cache line flush (see cache-v7.s) but that
would be difficult to maintain. So we are instead combining
the calls to dmac_clean_range() to amortize the cost.

This patch also removes setting the buffer type and buffer offset
for empty buffer sending since these fields are not read or needed
to process empty payloads.

Change-Id: Icc20a8ca86fc74b862470c840a197b7933536316
Signed-off-by: Selin Dag <sdag@codeaurora.org>